Create api_spec_doc
This is the full specification for TheWaypoint Management / Mooring Manager / Dock Manager API.

This API is meant to be used in conjunction with a Berth/Slip Management software, to accurately represent available inventory (berths) to TheWaypoint.

This API will place available slips to TheWaypoint and allow them to be booked and purchased online.  This API has the ability to place this inventory into multiple marketplaces (WebSites / APPS / Travel Agents / Mooring Brokers / Superyacht Agents).